requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.5
blinker==1.4
bz2file==0.98
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clize==4.1.1
cloudpickle==1.2.2
colorama==0.4.3
ConfigArgParse==1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
dopamine-rl==3.0.1
editdistance==0.5.3
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
future==0.18.2
fuzzywuzzy==0.18.0
gevent==1.4.0
gin-config==0.3.0
gitdb==4.0.2
GitPython==3.1.0
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.12.1
grpcio-tools==1.12.1
gunicorn==20.0.4
gym==0.15.4
h5py==2.10.0
httplib2==0.15.0
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
javalang==0.12.0
jedi==0.17.0
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kfac==0.2.0
kiwisolver==1.1.0
langdetect==1.0.8
Levenshtein-search==1.4.5
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mesh-tensorflow==0.1.9
mistune==0.8.4
more-itertools==8.0.2
mpmath==1.1.0
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.4.5
notebook==6.0.3
numba==0.47.0
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
od==1.0
opencv-python==4.1.2.30
OpenNMT-py==1.0.0
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pendulum==2.0.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
portalocker==1.5.2
prettytable==0.7.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
Pygments==2.6.1
pyonmttok==1.18.0
pyparsing==2.4.6
pypng==0.0.20
pyrsistent==0.16.0
pysolr==3.8.1
pytest==5.3.2
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
ray==0.8.4
redis==3.4.1
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
rouge==1.0.0
rsa==4.0
sacrebleu==1.4.3
scikit-learn==0.22.1
scipy==1.4.1
sea==2.1.0
seaborn==0.9.0
Send2Trash==1.5.0
sigtools==2.0.2
six==1.13.0
smmap==3.0.1
SoMaJo==2.0.4
soupsieve==2.0
subword-nmt==0.3.7
sympy==1.5.1
tabulate==0.8.7
tensor2tensor==1.15.2
tensorboard==2.1.0
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchtext==0.4.0
tornado==6.0.4
tqdm==4.30.0
traitlets==4.3.3
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==0.6.0
